Refactor CLI into package and add placeholder modules
