+++ cli/__init__.py
@@ -1,0 +1,5 @@
+"""SensibLaw command line interface package."""
+
+from .__main__ import main
+
+__all__ = ["main"]
