+++ src/frame/compiler.py
@@ -1,0 +1,14 @@
+"""Frame compilation utilities."""
+
+from __future__ import annotations
+
+
+def compile_frame(source: str) -> str:
+    """Compile a frame specification.
+
+    This placeholder implementation simply echoes the input.
+    """
+    return source
+
+
+__all__ = ["compile_frame"]
